+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/items/ItemSetting.java
@@ -139,15 +139,15 @@ public class ItemSetting<T> {
      */
     @SuppressWarnings("unchecked")
     public void load(@Nonnull SlimefunItem item) {
-        SlimefunPlugin.getItemCfg().setDefaultValue(item.getID() + '.' + getKey(), getDefaultValue());
-        Object configuredValue = SlimefunPlugin.getItemCfg().getValue(item.getID() + '.' + getKey());
+        SlimefunPlugin.getItemCfg().setDefaultValue(item.getId() + '.' + getKey(), getDefaultValue());
+        Object configuredValue = SlimefunPlugin.getItemCfg().getValue(item.getId() + '.' + getKey());
 
         if (defaultValue.getClass().isInstance(configuredValue)) {
             if (validateInput((T) configuredValue)) {
                 this.value = (T) configuredValue;
             } else {
                 Slimefun.getLogger().log(Level.WARNING, "Slimefun has found an invalid config setting in your Items.yml!");
-                Slimefun.getLogger().log(Level.WARNING, "  at \"{0}.{1}\"", new Object[] { item.getID(), getKey() });
+                Slimefun.getLogger().log(Level.WARNING, "  at \"{0}.{1}\"", new Object[] { item.getId(), getKey() });
                 Slimefun.getLogger().log(Level.WARNING, "{0} is not a valid input!", configuredValue);
                 Slimefun.getLogger().log(Level.WARNING, getErrorMessage());
             }
@@ -157,7 +157,7 @@ public class ItemSetting<T> {
 
             Slimefun.getLogger().log(Level.WARNING, "Slimefun has found an invalid config setting in your Items.yml!");
             Slimefun.getLogger().log(Level.WARNING, "Please only use settings that are valid.");
-            Slimefun.getLogger().log(Level.WARNING, "  at \"{0}.{1}\"", new Object[] { item.getID(), getKey() });
+            Slimefun.getLogger().log(Level.WARNING, "  at \"{0}.{1}\"", new Object[] { item.getId(), getKey() });
             Slimefun.getLogger().log(Level.WARNING, "Expected \"{0}\" but found: \"{1}\"", new Object[] { defaultValue.getClass().getSimpleName(), found });
         }
     }