+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/blocks/BlockPlacer.java
@@ -153,7 +153,7 @@ public class BlockPlacer extends SlimefunItem {
                     block.setType(item.getType());
                     block.getWorld().playEffect(block.getLocation(), Effect.STEP_SOUND, item.getType());
 
-                    BlockStorage.store(block, sfItem.getID());
+                    BlockStorage.store(block, sfItem.getId());
                     handler.onBlockPlacerPlace(e);
 
                     if (dispenser.getInventory().containsAtLeast(item, 2)) {
@@ -168,7 +168,7 @@ public class BlockPlacer extends SlimefunItem {
                 block.setType(item.getType());
                 block.getWorld().playEffect(block.getLocation(), Effect.STEP_SOUND, item.getType());
 
-                BlockStorage.store(block, sfItem.getID());
+                BlockStorage.store(block, sfItem.getId());
 
                 if (dispenser.getInventory().containsAtLeast(item, 2)) {
                     dispenser.getInventory().removeItem(new CustomItem(item, 1));