+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/magical/talismans/Talisman.java
@@ -79,7 +79,7 @@ public class Talisman extends SlimefunItem {
                 lore.add(line);
             }
 
-            enderTalisman = new SlimefunItemStack("ENDER_" + getID(), getItem().getType(), name, lore.toArray(new String[0]));
+            enderTalisman = new SlimefunItemStack("ENDER_" + getId(), getItem().getType(), name, lore.toArray(new String[0]));
         } else {
             enderTalisman = null;
         }
@@ -112,7 +112,7 @@ public class Talisman extends SlimefunItem {
     @Override
     public void postRegister() {
         EnderTalisman talisman = new EnderTalisman(this, getEnderVariant());
-        talisman.register(addon);
+        talisman.register(getAddon());
     }
 
     @Override