+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Added "slimefun.gps.bypass" permission to open GPS devices anywhere
 * (API) Added custom tags for developers
 * The range of the Seeker Pickaxe is now configurable
+* Added Energy Connector
 
 #### Changes
 * Improved Auto-Updater (Multi-Threading and more)
@@ -76,6 +77,10 @@
 * Fixed #2420
 * Fixed #2422
 * Fixed #2433
+* Fixed #2455
+* Fixed #2450
+* Fixed Steel Thrusters being used to milk cows
+* Fixed #2424
 
 ## Release Candidate 16 (07 Sep 2020)
 https://thebusybiscuit.github.io/builds/TheBusyBiscuit/Slimefun4/stable/#16